+++ application/common/components/Emailer.php
@@ -5,6 +5,7 @@ use common\helpers\MySqlDateTime;
 use common\models\EmailLog;
 use common\models\Method;
 use common\models\Mfa;
+use common\models\Password;
 use common\models\User;
 use InvalidArgumentException;
 use Psr\Log\LoggerInterface;
@@ -37,6 +38,9 @@ class Emailer extends Component
     const SUBJ_METHOD_REMINDER = 'REMINDER: Please verify your new password recovery method';
     const SUBJ_METHOD_PURGED = 'An unverified password recovery method has been removed from your {idpDisplayName}'
         . ' account';
+
+    const SUBJ_PASSWORD_EXPIRING = 'The password for your {idpDisplayName} Identity account is about to expire';
+    const SUBJ_PASSWORD_EXPIRED = 'The password for your {idpDisplayName} Identity account has expired';
 
     const PROP_SUBJECT = 'subject';
     const PROP_TO_ADDRESS = 'to_address';
@@ -88,6 +92,9 @@ class Emailer extends Component
     public $sendMethodReminderEmails = true;
     public $sendMethodPurgedEmails = true;
 
+    public $sendPasswordExpiringEmails = true;
+    public $sendPasswordExpiredEmails = true;
+
     /**
      * The list of subjects, keyed on message type. This is initialized during
      * the `init()` call during construction.
@@ -115,6 +122,9 @@ class Emailer extends Component
     public $subjectForMethodVerify;
     public $subjectForMethodReminder;
     public $subjectForMethodPurged;
+
+    public $subjectForPasswordExpiring;
+    public $subjectForPasswordExpired;
 
     /* The number of days of not using a security key after which we email the user */
     public $lostSecurityKeyEmailDays;
@@ -293,6 +303,9 @@ class Emailer extends Component
         $this->subjectForMethodReminder = $this->subjectForMethodReminder ?? self::SUBJ_METHOD_REMINDER;
         $this->subjectForMethodPurged = $this->subjectForMethodPurged ?? self::SUBJ_METHOD_PURGED;
 
+        $this->subjectForPasswordExpiring = $this->subjectForPasswordExpiring ?? self::SUBJ_PASSWORD_EXPIRING;
+        $this->subjectForPasswordExpired = $this->subjectForPasswordExpired ?? self::SUBJ_PASSWORD_EXPIRED;
+
         $this->subjects = [
             EmailLog::MESSAGE_TYPE_INVITE => $this->subjectForInvite,
             EmailLog::MESSAGE_TYPE_MFA_RATE_LIMIT => $this->subjectForMfaRateLimit,
@@ -310,6 +323,8 @@ class Emailer extends Component
             EmailLog::MESSAGE_TYPE_METHOD_PURGED => $this->subjectForMethodPurged,
             EmailLog::MESSAGE_TYPE_MFA_MANAGER => $this->subjectForMfaManager,
             EmailLog::MESSAGE_TYPE_MFA_MANAGER_HELP => $this->subjectForMfaManagerHelp,
+            EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRING => $this->subjectForPasswordExpiring,
+            EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRED => $this->subjectForPasswordExpired,
         ];
         
         $this->assertConfigIsValid();
@@ -641,6 +656,64 @@ class Emailer extends Component
 
         $this->logger->info([
             'action' => 'send method reminders',
+            'status' => 'finished',
+            'number_sent' => $numEmailsSent,
+        ]);
+    }
+
+    /**
+     * Iterates over all active users and sends email alert warning of impending password expiration
+     */
+    public function sendPasswordExpiringEmails()
+    {
+        if (! $this->sendPasswordExpiringEmails) {
+            return;
+        }
+
+        $numEmailsSent = 0;
+        $users = User::findAll(['active' => 'yes', 'locked' => 'no', ]);
+        foreach ($users as $user) {
+            /** @var Password $userPassword */
+            $userPassword = $user->currentPassword;
+            if (strtotime($userPassword->getExpiresOn()) < strtotime('+14 days') &&
+                ! $this->hasReceivedMessageRecently($user->id, EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRING)
+            ) {
+                $this->sendMessageTo(EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRING, $user);
+                $numEmailsSent++;
+            }
+        }
+
+        $this->logger->info([
+            'action' => 'send password expiring notices',
+            'status' => 'finished',
+            'number_sent' => $numEmailsSent,
+        ]);
+    }
+
+    /**
+     * Iterates over all active users and sends email alert warning of expired passwords
+     */
+    public function sendPasswordExpiredEmails()
+    {
+        if (! $this->sendPasswordExpiredEmails) {
+            return;
+        }
+
+        $numEmailsSent = 0;
+        $users = User::findAll(['active' => 'yes', 'locked' => 'no', ]);
+        foreach ($users as $user) {
+            /** @var Password $userPassword */
+            $userPassword = $user->currentPassword;
+            if (strtotime($userPassword->getExpiresOn()) < time() &&
+                ! $this->hasReceivedMessageRecently($user->id, EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRED)
+            ) {
+                $this->sendMessageTo(EmailLog::MESSAGE_TYPE_PASSWORD_EXPIRED, $user);
+                $numEmailsSent++;
+            }
+        }
+
+        $this->logger->info([
+            'action' => 'send password expired notices',
             'status' => 'finished',
             'number_sent' => $numEmailsSent,
         ]);